fixes #81 - completing update to WebDriver W3C Candidate Recommendation 26 September 2017
