+++ lib/WebDriver/Frame.php
@@ -30,6 +30,8 @@ namespace WebDriver;
  */
 final class Frame extends AbstractWebDriver
 {
+    const WEBDRIVER_FRAME_ID = 'frame-075b-4da1-b6ba-e579c2d3230a';
+
     /**
      * {@inheritdoc}
      */
